[web_server] move v1 code to separate file (#7091)
